Extract shared duty expression prettifier
