+++ app/formatters/duty_expression_formatter.rb
@@ -1,14 +1,7 @@
 class DutyExpressionFormatter
-  class << self
-    def prettify(float)
-      TradeTariffBackend.number_formatter.number_with_precision(
-        float,
-        minimum_decimal_points: 2,
-        precision: 4,
-        strip_insignificant_zeros: true,
-      )
-    end
+  extend DutyExpressionPrettifier
 
+  class << self
     def format(opts = {})
       context = Context.build(opts)
       output = OutputBuilder.call(context)
