+++ app/formatters/requirement_duty_expression_formatter.rb
@@ -1,14 +1,7 @@
 class RequirementDutyExpressionFormatter
-  class << self
-    def prettify(float)
-      TradeTariffBackend.number_formatter.number_with_precision(
-        float,
-        precision: 4,
-        minimum_decimal_points: 2,
-        strip_insignificant_zeros: true,
-      )
-    end
+  extend DutyExpressionPrettifier
 
+  class << self
     def format(opts = {})
       context = Context.build(opts)
